+++ composeApp/src/commonMain/kotlin/tech/kotlinlang/permission/Platform.kt
@@ -1,7 +1,0 @@
-package tech.kotlinlang.permission
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform